Refactor some messages into `introducer_protocol`.
